requirements.txt
aiohttp
python-dotenv
cryptofeed
aiokafka
asyncio
streamlit
kafka-python
